// apps/desktop/src/ipc/generated/PackageReviewSummary.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type PackageReviewSummary = { packageId: string, projectTitle: string, draftName: string, exportedBy: string, exportedAt: number, myCommentCount: number, responseExportedAt: number | null, };